# packages/contracts/openapi/careos-api.v1.json
{
  "components": {
    "schemas": {
...
      "ChangeRoleRequest": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "role"
        ],
        "title": "ChangeRoleRequest",
        "type": "object"
      },
      "CloseIncidentRequest": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "CloseIncidentRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentDetail": {
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "allowed_actions": {
            "description": "Actions the current user may take now: takeover, resolve, close.",
            "items": {
              "type": "string"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "closed_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "escalation": {
            "items": {
              "$ref": "#/components/schemas/ScheduledActionView"
            },
            "title": "Escalation",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationView"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NextActionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/IncidentPriority"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "resolution_category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "service_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceUserBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "trigger_type": {
            "title": "Trigger Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reference",
          "status",
          "priority",
          "trigger_type",
          "is_active",
          "created_at",
          "updated_at",
          "acknowledged_at",
          "resolved_at",
          "closed_at",
          "service_user",
          "device",
          "assignee",
          "next_action",
          "location",
          "resolution_category",
          "resolution_notes",
          "resolved_by",
          "closed_by",
          "escalation",
          "allowed_actions"
        ],
        "title": "IncidentDetail",
        "type": "object"
      },
...
      "IncidentPriority": {
        "enum": [
          "CRITICAL",
          "HIGH",
          "MEDIUM",
          "LOW"
        ],
        "title": "IncidentPriority",
        "type": "string"
      },
...
      "IncidentSummary": {
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "next_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NextActionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/IncidentPriority"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "service_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceUserBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "trigger_type": {
            "title": "Trigger Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reference",
          "status",
          "priority",
          "trigger_type",
          "is_active",
          "created_at",
          "updated_at",
          "acknowledged_at",
          "resolved_at",
          "closed_at",
          "service_user",
          "device",
          "assignee",
          "next_action"
        ],
        "title": "IncidentSummary",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "ResolveIncidentRequest": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ResolutionCategory"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "category"
        ],
        "title": "ResolveIncidentRequest",
        "type": "object"
      },
...
      "SessionView": {
        "properties": {
          "csrf_token": {
            "description": "Send as X-CSRF-Token on every state-changing request.",
            "title": "Csrf Token",
            "type": "string"
          },
          "organisation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrganisationView"
              },
              {
                "type": "null"
              }
            ]
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/UserView"
          }
        },
        "required": [
          "user",
          "organisation",
          "permissions",
          "csrf_token"
        ],
        "title": "SessionView",
        "type": "object"
      },
...
      "UserView": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "organisation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organisation Id"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "organisation_id",
          "is_active",
          "last_login_at"
        ],
        "title": "UserView",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CSRF": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "GatewayKey": {
        "in": "header",
        "name": "X-CareOS-Gateway-Key",
        "type": "apiKey"
      },
      "Session": {
        "in": "cookie",
        "name": "careos_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "CareOS does not diagnose, predict illness or recommend treatment. AI features are assistive only; incident handling is deterministic.",
    "summary": "Telecare & safety incident orchestration platform",
    "title": "CareOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "me_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/incidents": {
      "get": {
        "operationId": "list_incidents_v1_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "awaiting_closure",
                "recent"
              ],
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentSummary"
                  },
                  "title": "Response List Incidents V1 Incidents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "List Incidents",
        "tags": [
          "incidents"
        ]
      }
    },
    "/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident_v1_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Get Incident",
        "tags": [
          "incidents"
        ]
      }
    },
    "/v1/incidents/{incident_id}/close": {
      "post": {
        "operationId": "close_v1_incidents__incident_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseIncidentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Close",
        "tags": [
          "incidents"
        ]
      }
    },
    "/v1/incidents/{incident_id}/resolve": {
      "post": {
        "operationId": "resolve_v1_incidents__incident_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveIncidentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Resolve",
        "tags": [
          "incidents"
        ]
      }
    },
    "/v1/incidents/{incident_id}/takeover": {
      "post": {
        "operationId": "take_over_v1_incidents__incident_id__takeover_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Take Over",
        "tags": [
          "incidents"
        ]
      }
    },
    "/v1/incidents/{incident_id}/timeline": {
      "get": {
        "operationId": "get_timeline_v1_incidents__incident_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentEventView"
                  },
                  "title": "Response Get Timeline V1 Incidents  Incident Id  Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Get Timeline",
        "tags": [
          "incidents"
        ]
      }
    },
...
    "/v1/users": {
      "get": {
        "operationId": "list_users_v1_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserView"
                  },
                  "title": "Response List Users V1 Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "create_user_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Create User",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/{user_id}/role": {
      "patch": {
        "operationId": "change_role_v1_users__user_id__role_patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Change Role",
        "tags": [
          "users"
        ]
      }
    }
  }
}